% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_barchart.R
\name{radial_barchart_static}
\alias{radial_barchart_static}
\title{Creates a radial barchart which shows a cluster's average feature values.}
\usage{
radial_barchart_static(
  df,
  group_names,
  inner_label = "",
  color_inner_circle = "grey60",
  scale_rng = c(-1, 1) * 1.5,
  interactive = FALSE,
  tooltip_bars = "",
  tooltip_labels = NULL,
  show_group_names = TRUE
)
}
\arguments{
\item{df}{The dataset as data frame. Should be scaled. See ?scale.
Should only contain the category values in a wide table form.}

\item{group_names}{Character vector with group names of features.
The number of columns in df and the length of group_names must be equal. Group
names are displayed in the inner circle.}

\item{inner_label}{Label inside the inner circle, e.g. name of the chart (optional).}

\item{color_inner_circle}{Color of the inner circle (optional).}

\item{scale_rng}{Min and max values to be shown (optional).}

\item{interactive}{Boolean to indicate whether the plot should be interactive or not (optional).}

\item{tooltip_bars}{String to choose which tooltip texts will be displayed for the bars (when interactive).
Options to select from: average, sd (standard deviation), ci (confidence interval), all. Default is average (optional).}

\item{tooltip_labels}{Data frame to provide the tooltip texts for the labels (when interactive), needs to have a "label" and a "description" column. The description provides the tooltip text for the assigned label (optional).}

\item{show_group_names}{Boolean to indicate whether the group names should be shown inside the inner circle. Default is true (optional).}
}
\value{
the radial barchart as ggplot
}
\description{
Creates a radial barchart which shows a cluster's average feature values.
}
\examples{
radial_barchart_static(data_static, group_names, inner_label = "Results", color_inner_circle = "#E3E3E3", scale_rng = c(-1.6,1.6), interactive = TRUE, tooltip_bars = "sd", tooltip_labels = data.frame(label = c("f_2"), description = c("Mock Description")), show_group_names = FALSE)
}
