% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_linechart.R
\name{radial_line_chart}
\alias{radial_line_chart}
\title{Creates a radial line chart which shows the average feature values for each cluster.}
\usage{
radial_line_chart(
  df,
  group_names,
  cluster = ".cluster",
  inner_label = "",
  color_inner_circle = "grey60",
  color_clusters = NULL,
  scale_rng = c(-1, 1) * 1.5,
  interactive = FALSE,
  tooltip_bars = "",
  tooltip_labels = NULL,
  show_group_names = TRUE
)
}
\arguments{
\item{df}{The data set as data frame. Should be scaled. See ?scale.
Should only contain the category values in a wide table form.}

\item{group_names}{Character vector with group names of features.
The number of columns in df and the length of group_names must be equal. Group
names are displayed in the inner circle.}

\item{cluster}{String containing the column name, which indicates the cluster membership in df (optional).}

\item{inner_label}{Label inside the inner circle, e.g. name of the chart (optional).}

\item{color_inner_circle}{Color of the inner circle (optional).}

\item{color_clusters}{Colors of the different clusters (optional).}

\item{scale_rng}{Min and max values to be shown (optional).}

\item{interactive}{Boolean to indicate whether the plot should be interactive or not (optional).}

\item{tooltip_bars}{String to choose which tooltip texts will be displayed for the bars (when interactive).
Options to select from: average, sd (standard deviation), ci (confidence interval), all. Default is average (optional).}

\item{tooltip_labels}{Data frame to provide the tooltip texts for the labels (when interactive), needs to have a "label" and a "description" column. The description provides the tooltip text for the assigned label (optional).}

\item{show_group_names}{Boolean to indicate whether the group names should be shown inside the inner circle. Default is true (optional).}
}
\value{
the radial line chart as ggplot
}
\description{
Creates a radial line chart which shows the average feature values for each cluster.
}
